Remove sync.WaitGroup from lobby client
Much like the excessive concurrency removed in 2f57dab2f38e4efc032, this commit
removes a WaitGroup with a single member in favor of a blocking channel read.

In place it exposes a 'Done' method on the client, which blocks on the internal
'exit' channel.

Kudos to dgryski for suggesting this simplification in review.