Client 'send' no longer returns an error
The only response to client send errors is to log, so log once in the client
and reduce the noise higher up in the program. There's a bit more context
available at the higher level, but the contents of the command and prepared
payload should be enough to make clear what was going on.

In terms of error handling, 'send' already results in the connection closing,
which gracefully resets all client/matchbot state.