struct-ify game, split up matchbot code
This commit changes game.Start from a single function to a struct with various
methods. This will greatly simplify game management/reporting in a central
location (like a web UI). In the process I tried to make game.Game a little
better contained (e.g., don't ship it a lobby client just so it can call
ConnectUser one time).

This also splits out the game-related functionality of the matchbot into
matchbot/game.go, as matchbot/matchbot.go was getting a bit large. the Matchbot
might be a little too much "all seeing all knowing", but it's a handy design
for now.